refactor - alterado o tipo dos números na entidade e no ddl

+++ src/main/java/com/jazztech/creditanalysis/presentation/dto/ResponseDto.java
@@ -1,6 +1,7 @@
 package com.jazztech.creditanalysis.presentation.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -8,9 +9,9 @@ import java.util.UUID;
 public record ResponseDto(
         UUID id,
         Boolean approved,
-        Double approvedLimit,
-        Double withdraw,
-        Double annualInterest,
+        BigDecimal approvedLimit,
+        BigDecimal withdraw,
+        BigDecimal annualInterest,
         UUID clientId,
         LocalDateTime date
 ) {
